# [Julia] Remove StaticArrays.jl, replace with Vector5 types