--- a/julia/Project.toml
+++ b/julia/Project.toml
@@ -1,0 +1,5 @@
+[deps]
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"