--- a/julia_con/related.jl
+++ b/julia_con/related.jl
@@ -1,12 +1,16 @@
+using Pkg
+Pkg.instantiate()
+
 using Base.Threads
 
 using JSON3
 using StructTypes
 using Dates
-using StaticArrays
 using ChunkSplitters
 
 # warmup is done by hyperfine
+
+include("../julia/vector5.jl")
 
 function relatedIO()
     json_string = read("../posts.json", String)
@@ -33,7 +37,7 @@ end
 struct RelatedPost
     _id::String
     tags::Vector{String}
-    related::SVector{5,PostData}
+    related::SVector5{PostData}
 end
 
 StructTypes.StructType(::Type{PostData}) = StructTypes.Struct()
@@ -105,7 +109,7 @@ function related(::Type{T}, posts) where {T}
 
             fastmaxindex!(taggedpostcount, topn, maxn, maxv)
 
-            relatedpost = RelatedPost(post._id, post.tags, SVector{topn}(@view posts[maxn]))
+            relatedpost = RelatedPost(post._id, post.tags, SVector5(@view posts[maxn]))
             relatedposts[i] = relatedpost
         end
     end