--- a/julia_con/related.jl
+++ b/julia_con/related.jl
@@ -1,6 +1,3 @@
-using Pkg
-Pkg.instantiate()
-
 using Base.Threads
 
 using JSON3