This entirely removes StaticArrays.jl from the Julia code. Instead, we implement new {M,S}Vector5 types specialized for mutable or immutable 5-element collections.

Also, this replaces manual package management with Project.toml files including the package UUIDs, adding specificity and additional security.

Fixes #194