@jinyus 2023-10-12
I not sure how I feel about this. Making a specialized, fixed size array definitely goes against the spirit of the competition and other solutions will mimic this to level the playing field.

With that said, I'm not seeing a speed up on my end. This and the old solution is pretty much equal(linux). How are the results looking for you?

old:

```
Benchmark 1: julia related.jl
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 21 milliseconds
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 22 milliseconds
Processing time (w/o IO): 22 milliseconds
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 22 milliseconds
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 22 milliseconds
Processing time (w/o IO): 22 milliseconds
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 22 milliseconds
Processing time (w/o IO): 22 milliseconds
  Time (mean ± σ):      2.011 s ±  0.019 s    [User: 1.860 s, System: 0.143 s]
  Range (min … max):    1.966 s …  2.033 s    10 runs
```  

new:

```
Benchmark 1: julia --project related.jl
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 22 milliseconds
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 22 milliseconds
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 22 milliseconds
Processing time (w/o IO): 23 milliseconds
Processing time (w/o IO): 22 milliseconds
Processing time (w/o IO): 24 milliseconds
Processing time (w/o IO): 22 milliseconds
  Time (mean ± σ):      1.728 s ±  0.024 s    [User: 1.569 s, System: 0.150 s]
  Range (min … max):    1.700 s …  1.767 s    10 runs
```