@mkitti 2023-10-12
I'm running on a potato at the moment, but I do see an impact versus main.

Old:
```
Benchmark 1: julia related.jl
Processing time (w/o IO): 57 milliseconds
Processing time (w/o IO): 68 milliseconds
Processing time (w/o IO): 57 milliseconds
Processing time (w/o IO): 68 milliseconds
Processing time (w/o IO): 55 milliseconds
Processing time (w/o IO): 57 milliseconds
Processing time (w/o IO): 67 milliseconds
Processing time (w/o IO): 68 milliseconds
Processing time (w/o IO): 68 milliseconds
Processing time (w/o IO): 57 milliseconds
Processing time (w/o IO): 68 milliseconds
Processing time (w/o IO): 69 milliseconds
Processing time (w/o IO): 56 milliseconds
  Time (mean ± σ):      4.232 s ±  0.018 s    [User: 4.037 s, System: 0.380 s]
  Range (min … max):    4.204 s …  4.262 s    10 runs
```

New:
```
Benchmark 1: julia --project related.jl
Processing time (w/o IO): 49 milliseconds
Processing time (w/o IO): 48 milliseconds
Processing time (w/o IO): 48 milliseconds
Processing time (w/o IO): 49 milliseconds
Processing time (w/o IO): 49 milliseconds
Processing time (w/o IO): 49 milliseconds
Processing time (w/o IO): 48 milliseconds
Processing time (w/o IO): 48 milliseconds
Processing time (w/o IO): 50 milliseconds
Processing time (w/o IO): 49 milliseconds
Processing time (w/o IO): 50 milliseconds
Processing time (w/o IO): 49 milliseconds
Processing time (w/o IO): 49 milliseconds
  Time (mean ± σ):      3.634 s ±  0.078 s    [User: 3.480 s, System: 0.339 s]
  Range (min … max):    3.551 s …  3.795 s    10 runs
```