@jinyus 2023-10-12
Just tested it on the benchmarking VM. Slower at 5k, same at 20k and a tap slower at 60k:

old:

```
5k -> 34ms
20k -> 576ms
60k -> 5.04s

```

new:

```
Julia:

        Benchmark 1: julia --project related.jl
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 38 milliseconds
          Time (mean ± σ):      2.800 s ±  0.030 s    [User: 2.622 s, System: 0.177 s]
          Range (min … max):    2.771 s …  2.861 s    10 runs
         
Julia:

        Benchmark 1: julia --project related.jl
        Processing time (w/o IO): 577 milliseconds
        Processing time (w/o IO): 577 milliseconds
        Processing time (w/o IO): 577 milliseconds
          Time (mean ± σ):      4.009 s ±  0.004 s    [User: 3.822 s, System: 0.186 s]
          Range (min … max):    4.007 s …  4.012 s    2 runs
         
Julia:

        Benchmark 1: julia --project related.jl
        Processing time (w/o IO): 5077 milliseconds
        Processing time (w/o IO): 5066 milliseconds
        Processing time (w/o IO): 5058 milliseconds
          Time (mean ± σ):     13.344 s ±  0.003 s    [User: 13.052 s, System: 0.282 s]
          Range (min … max):   13.342 s … 13.346 s    2 runs

```