@mkitti 2023-10-12
Hmm... I guess not much point in terms of performance then. Let's hold on for the moment. I see some things to try.