@MasonProtter 2023-10-12
Why not just use a `NTuple`?