@mkitti 2023-10-12
For the SVector5 implementation, sure, but we also need a `MVector` replacement.