@MasonProtter 2023-10-12
Hm, nevermind, after doing some benchmarks on my machine, the above suggestion I posted is actually slower. I also observe @mkitti's implementation here recovering most of the performance lost when StaticArrays was dropped.