@mkitti 2023-10-12
I had tried a version of this where I inline the call to get/setfield and things seem to get worse.

The problem is that the field functions have to do another check to validate the input and possibly throw. I think our branching ends up being cheaper.

I have also prepared a package refactoring to take advantage of precompile disk caching in the form of pkgimages:

https://github.com/mkitti/related_post_gen/pull/1