@algunion 2023-10-12
> I not sure how I feel about this. Making a specialized, fixed size array definitely goes against the spirit of the competition and other solutions will mimic this to level the playing field.

@jinyus,

Consider this: we already have this in the Julia ecosystem (e.g., `StaticArrays.jl`). The package is blazing fast and safe (take a look at [arguments ](https://github.com/jinyus/related_post_gen/issues/194#issuecomment-1759217426)like this one).

Now, you are saying that if @mkitti implements his code in this repo, this feels like going against the spirit of the competition. But if we go and publish it on the Julia registry (and it becomes something more general like having a Julia macro that generates specialized code for any small N - 5 included - and we call it as a package, then we are OK).

Also, if people are to invest more time in this benchmark, I think we need to have some transparent and predictable rules and also make sure those rules are consistently applied to all languages involved. See [here](https://github.com/jinyus/related_post_gen/issues/203).

Now it seems a new rule appeared: if you do not have a package that does X, you are also not allowed to implement X - even if you do it in a transparent and clearly safe way.