@MasonProtter 2023-10-12
Maybe superseded by https://github.com/jinyus/related_post_gen/pull/207 which is a more general approach and doesn't require hand-made structs of a specific size.