@dlakelan 2023-10-12
The reason for this is that StaticArrays.jl was ruled as breaking the contests rules because it "invokes unsafe code". So this is just a reimplementation of the subset of functionality needed from StaticArrays but without invoking any "unsafe" code...

for those who haven't been following the relevant discussion is here: 

https://github.com/jinyus/related_post_gen/issues/194